Add files via upload
This code defines a function called deleteNote that expects one noteId argument. The function uses the fetch method to send the request to the server. The fetch method returns a Promise that will be resolved when a response is returned to the request.

fetch expects two arguments: the first argument is the URL to send the request to, and the second argument is an object with the request configuration. The configuration object has two entries in this code:

    method: specifies the HTTP method to be used for the request. In this case, the POST method is used.
    points: specifies the content of the request. In this case, a single-item JSON object is used, where the key is "noteId" and the value is the noteId that was passed as an argument to the deleteNote function.

After sending a request to the server using fetch, the then method is called on the returned Promise, which accepts a function as an argument. This function will be executed when a response is returned to the request. In this function, a new URL is set using window.location.href, which redirects the user to the home page.